+++ packages/terra-image/CHANGELOG.md
@@ -8,6 +8,10 @@ Unreleased
 * Added `object-fit` property in css.
 * Added prop Fit.
 * Added examples using Fit.
+* Added documentation on required peerDependencies
+
+### Changed
+* Update tests to use `Terra.describeViewports` for setting viewports and use `Terra.it.validatesElement`
 
 3.11.0  - (July 11, 2019)
 ------------------